pyproject.toml
[project]
name = "ragbase"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = ["fastembed>=0.7.0", "flashrank>=0.2.10", "langchain-community>=0.3.24", "langchain-experimental>=0.3.4", "langchain-openai>=0.3.17", "langchain-qdrant>=0.2.0", "langchain-core>=0.3.60", "python-dotenv>=1.1.0", "qdrant-client>=1.14.2", "streamlit>=1.45.1", "pypdfium2>=4.30.1", "langchain-groq>=0.3.2"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.build.pdm]
package-dir='src'

[tool.pdm]
distribution = false
